requirements.txt
pydantic-settings==2.12.0
fastapi==0.127.1
uvicorn==0.40.0
sqlalchemy==2.0.45
pydantic==2.12.5
requests==2.32.5
asyncpg==0.31.0
aioboto3==15.5.0
python-multipart==0.0.21
openai==2.14.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.21
pydantic[email]==2.12.5
bcrypt==4.0.1
authlib==1.6.6
httpx==0.28.1
itsdangerous==2.2.0